fix: Bug rpm của súng từ số nguyên sang số thực

+++ src/GameClient/Assets/Scripts/Guns/RifleGun.cs
@@ -10,7 +10,7 @@ public class RifleGun : MonoBehaviour
 {
     public UnityEvent OnShoot;
     
-    private const int SECONDS_PER_MINUTE = 60;
+    private const float SECONDS_PER_MINUTE = 60f;
     private readonly int FireStateAnimHash = Animator.StringToHash("AlternateSingleFire");
 
     [SerializeField]
